azure-event-hub-master/service/requirements.txt
azure-eventhub==1.3.1
Flask==1.0.2
CherryPy==18.0.0